requirements.txt
python-dotenv>=1.0.0
schedule>=1.2.0
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
twilio>=8.10.0
gunicorn==20.1.0
